@@ -32,32 +27,32 @@
             <organizationUrl>https://github.com/mosip/mimoto</organizationUrl>
         </developer>
     </developers>
-    <properties>
-        <java.version>11</java.version>
 
+    <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
             <!-- mosip -->
-        <mosip.kernel.core.version>1.2.0.1-B1</mosip.kernel.core.version>
-        <mosip.biometric.util.version>1.2.0.1-B1</mosip.biometric.util.version>
-        <mosip.kernel.auth.adapter.version>1.2.0.1-B1</mosip.kernel.auth.adapter.version>
-        <kernel.biometrics.api.version>1.2.0.1-B2</kernel.biometrics.api.version>
-        <kernel.cbeffutil.version>1.2.0.1-B1</kernel.cbeffutil.version>
+        <mosip.kernel.core.version>1.2.1-java21-SNAPSHOT</mosip.kernel.core.version>
+        <mosip.biometric.util.version>1.2.1-java21-SNAPSHOT</mosip.biometric.util.version>
+        <mosip.kernel.auth.adapter.version>1.2.1-java21-SNAPSHOT</mosip.kernel.auth.adapter.version>
+        <kernel.biometrics.api.version>1.2.1-java21-SNAPSHOT</kernel.biometrics.api.version>
+        <kernel.cbeffutil.version>1.2.1-java21-SNAPSHOT</kernel.cbeffutil.version>
+
         <vertx.version>3.9.1</vertx.version>
 
             <!-- maven -->
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
         <maven.compiler.version>3.8.0</maven.compiler.version>
         <maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
         <maven.war.plugin.version>3.1.0</maven.war.plugin.version>
         <maven.surefire.plugin.version>2.22.0</maven.surefire.plugin.version>
-        <maven.jacoco.version>0.8.5</maven.jacoco.version>
+        <maven.jacoco.version>0.8.11 </maven.jacoco.version>
         <maven.sonar.plugin.version>3.7.0.1746</maven.sonar.plugin.version>
-        <maven.javadoc.version>3.0.1</maven.javadoc.version>
+        <maven.javadoc.version>3.6.3</maven.javadoc.version>
 
             <!-- spring -->
-        <spring.boot.version>2.0.2.RELEASE</spring.boot.version>
+        <spring.boot.version>3.2.3</spring.boot.version>
 
 
             <!-- data -->
