@@ -1,5 +1,7 @@
 package io.mosip.mimoto.config;
 
+import io.mosip.kernel.auth.defaultadapter.filter.AuthFilter;
+import io.mosip.kernel.auth.defaultadapter.filter.CorsFilter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -7,24 +9,26 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
-import io.mosip.kernel.auth.defaultadapter.filter.AuthFilter;
-import io.mosip.kernel.auth.defaultadapter.filter.CorsFilter;
 import io.mosip.kernel.core.logger.spi.Logger;
 import io.mosip.mimoto.util.LoggerUtil;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.annotation.web.configurers.ExceptionHandlingConfigurer;
+import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
 
 import java.util.HashMap;
 import java.util.Map;
 
 @Configuration
 @EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
+@EnableMethodSecurity
 @Order(1)
-public class Config extends WebSecurityConfigurerAdapter {
+public class Config {
     private Logger logger = LoggerUtil.getLogger(Config.class);
 
     @Value("${mosipbox.public.url}")
@@ -45,28 +49,31 @@ public class Config extends WebSecurityConfigurerAdapter {
         return new HashMap<>();
     }
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-
-        if (!isCSRFEnable) {
-            http = http.csrf().disable();
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        if (isCSRFEnable) {
+            http.csrf(AbstractHttpConfigurer::disable);
         }
 
-        // FIXME: web security is disabled for all enpoints
-        http.authorizeRequests().antMatchers("*").authenticated().anyRequest().permitAll().and().exceptionHandling();
+        //TODO: exceptionHandling left
+        http.authorizeHttpRequests(authorizeRequests -> authorizeRequests
+                .requestMatchers("*")
+                .authenticated()
+                .anyRequest()
+                .permitAll()
+        );
+        http.exceptionHandling(exceptionConfigurer  -> new ExceptionHandlingConfigurer());
+
         if (isCORSEnable) {
             http.addFilterBefore(new CorsFilter(origins), AuthFilter.class);
         }
-        http.headers().cacheControl();
-        http.headers().frameOptions().sameOrigin();
+        http.headers(headersEntry -> {
+            headersEntry.cacheControl(Customizer.withDefaults());
+            headersEntry.frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin);
+        });
+
+        return http.build();
 
     }
 
-    @Bean
-    public TaskScheduler taskScheduler() {
-        ThreadPoolTaskScheduler threadPoolTaskScheduler = new ThreadPoolTaskScheduler();
-        threadPoolTaskScheduler.setPoolSize(5);
-        threadPoolTaskScheduler.setThreadNamePrefix("ThreadPoolTaskScheduler");
-        return threadPoolTaskScheduler;
-    }
 }
