@@ -1,7 +1,7 @@
 package io.mosip.mimoto.core.http;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 import lombok.Data;
 
