@@ -3,8 +3,8 @@ package io.mosip.mimoto.dto;
 import com.google.gson.annotations.Expose;
 import lombok.Data;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
 
 @Data
 public class DisplayDTO {
