@@ -3,7 +3,7 @@ package io.mosip.mimoto.dto;
 import com.google.gson.annotations.Expose;
 import lombok.Data;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.List;
 
 @Data
