@@ -2,7 +2,7 @@ package io.mosip.mimoto.dto.mimoto;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @Data
 public class VCCredentialResponseProof {
