@@ -1,16 +1,27 @@
 package io.mosip.mimoto.config;
 
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.ssl.TrustStrategy;
+import org.apache.hc.client5.http.classic.HttpClient;
+import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
+import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManager;
+import org.apache.hc.client5.http.socket.ConnectionSocketFactory;
+import org.apache.hc.client5.http.socket.PlainConnectionSocketFactory;
+import org.apache.hc.client5.http.ssl.SSLConnectionSocketFactory;
+import org.apache.hc.core5.http.URIScheme;
+import org.apache.hc.core5.http.config.Registry;
+import org.apache.hc.core5.http.config.RegistryBuilder;
+import org.apache.hc.core5.ssl.SSLContextBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
+import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.annotation.web.configurers.ExceptionHandlingConfigurer;
+import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.web.client.RestTemplate;
 
 import javax.net.ssl.SSLContext;
@@ -18,40 +29,65 @@ import java.security.KeyManagementException;
 import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.X509Certificate;
-import java.util.HashMap;
-import java.util.Map;
+
 
 @Configuration
 @ComponentScan(basePackages = {
         "io.mosip.mimoto.*" }, excludeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, value = {
         Config.class, OpenApiProperties.class, SwaggerConfig.class}))
-public class MimotoConfig extends WebSecurityConfigurerAdapter {
+public class MimotoConfig {
 
-    @Override
+   /* @Override
     protected void configure(HttpSecurity http) throws Exception {
         http = http.csrf().disable();
         http.authorizeRequests().antMatchers("*").authenticated().anyRequest().permitAll().and().exceptionHandling();
         http.headers().cacheControl();
         http.headers().frameOptions().sameOrigin();
 
+    }*/
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http.csrf(AbstractHttpConfigurer::disable);
+        http.authorizeHttpRequests(authorizeRequests -> authorizeRequests
+                .requestMatchers("*")
+                .authenticated()
+                .anyRequest()
+                .permitAll()
+
+        );
+        http.exceptionHandling(exceptionConfigurer  -> new ExceptionHandlingConfigurer());
+        http.headers(headersEntry -> {
+            headersEntry.cacheControl(Customizer.withDefaults());
+            headersEntry.frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin);
+        });
+
+        return http.build();
     }
 
     @Bean
     public RestTemplate selfTokenRestTemplate()
             throws KeyManagementException, NoSuchAlgorithmException, KeyStoreException {
 
-        TrustStrategy acceptingTrustStrategy = (X509Certificate[] chain, String authType) -> true;
+        final RestTemplate restTemplate = new RestTemplate();
 
-        SSLContext sslContext = org.apache.http.ssl.SSLContexts.custom().loadTrustMaterial(null, acceptingTrustStrategy)
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadTrustMaterial((X509Certificate[] certificateChain, String authType) -> true)  // <--- accepts each certificate
                 .build();
 
-        SSLConnectionSocketFactory csf = new SSLConnectionSocketFactory(sslContext);
+        Registry<ConnectionSocketFactory> socketRegistry = RegistryBuilder.<ConnectionSocketFactory>create()
+                .register(URIScheme.HTTPS.getId(), new SSLConnectionSocketFactory(sslContext))
+                .register(URIScheme.HTTP.getId(), new PlainConnectionSocketFactory())
+                .build();
 
-        CloseableHttpClient httpClient = HttpClients.custom().setSSLSocketFactory(csf).build();
-        HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory();
+        HttpClient httpClient = HttpClientBuilder.create()
+                .setConnectionManager(new PoolingHttpClientConnectionManager(socketRegistry))
+                .setConnectionManagerShared(true)
+                .build();
 
-        requestFactory.setHttpClient(httpClient);
-        return new RestTemplate(requestFactory);
+        ClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);
+        restTemplate.setRequestFactory(requestFactory);
+        return restTemplate;
 
     }
 
@@ -59,24 +95,32 @@ public class MimotoConfig extends WebSecurityConfigurerAdapter {
     public RestTemplate plainRestTemplate()
             throws KeyManagementException, NoSuchAlgorithmException, KeyStoreException {
 
-        TrustStrategy acceptingTrustStrategy = (X509Certificate[] chain, String authType) -> true;
+        final RestTemplate restTemplate = new RestTemplate();
 
-        SSLContext sslContext = org.apache.http.ssl.SSLContexts.custom().loadTrustMaterial(null, acceptingTrustStrategy)
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadTrustMaterial((X509Certificate[] certificateChain, String authType) -> true)  // <--- accepts each certificate
                 .build();
 
-        SSLConnectionSocketFactory csf = new SSLConnectionSocketFactory(sslContext);
+        Registry<ConnectionSocketFactory> socketRegistry = RegistryBuilder.<ConnectionSocketFactory>create()
+                .register(URIScheme.HTTPS.getId(), new SSLConnectionSocketFactory(sslContext))
+                .register(URIScheme.HTTP.getId(), new PlainConnectionSocketFactory())
+                .build();
 
-        CloseableHttpClient httpClient = HttpClients.custom().setSSLSocketFactory(csf).build();
-        HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory();
+        HttpClient httpClient = HttpClientBuilder.create()
+                .setConnectionManager(new PoolingHttpClientConnectionManager(socketRegistry))
+                .setConnectionManagerShared(true)
+                .build();
 
-        requestFactory.setHttpClient(httpClient);
-        return new RestTemplate(requestFactory);
+        ClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);
+        restTemplate.setRequestFactory(requestFactory);
+        return restTemplate;
 
     }
 
+/*
 
     @Bean
     public Map<String, String> injiConfig() {
         return new HashMap<>();
-    }
+    }*/
 }
