@@ -47,7 +47,7 @@ public class IssuersController {
     private final Logger logger = LoggerFactory.getLogger(IssuersController.class);
 
     @GetMapping()
-    public ResponseEntity<Object> getAllIssuers(@RequestParam(required = false) String search) {
+    public ResponseEntity<Object> getAllIssuers(@RequestParam(required = false, name = "search") String search) {
         ResponseWrapper<IssuersDTO> responseWrapper = new ResponseWrapper<>();
         responseWrapper.setId(ID);
         responseWrapper.setVersion("v1");
@@ -92,7 +92,7 @@ public class IssuersController {
 
     @GetMapping("/{issuer-id}/credentialTypes")
     public ResponseEntity<Object> getCredentialTypes(@PathVariable("issuer-id") String issuerId,
-                                                                   @RequestParam(required = false) String search) {
+                                                                   @RequestParam(required = false, name = "search") String search) {
         ResponseWrapper<Object> responseWrapper = new ResponseWrapper<>();
         responseWrapper.setId(ID);
         responseWrapper.setVersion("v1");
